[6.1] Align SqlException Numbers across platforms (#3475)
* Reduce automated test crashes (#2968)

* Converted Threads to long-running Tasks

The key advantage is that exceptions propagate properly.
If a thread throws an exception (as a result of a failed test assertion, or otherwise) then the test host crashes and must be restarted.

* Corrected the instantiation of the cancellation task - missing state parameter.

* Changes to TestSqlCommandCancel, eliminating timing-specific cancellation behaviour testing.
This should also allow the test to run on both netcore and netfx.

* Responding to code review.
* Removed two unnecessary iterations from DatabaseHelper.
* Added explanatory comments to ApiShould.
* Switched to using Task.WaitAll rather than waiting for each Task in sequence.

* Improve cancellation detection

Cancellation can trigger one of several different errors, resulting in a flakier test.
Also ensure that the query always takes more than 150ms, ensuring that a quick query execution doesn't cause the test to fail.
Finally, make sure that we try to read everything from the SqlDataReader.

* Correcting previous merge

* Align SqlException Numbers across platforms (#3461)

* - Align SqlException Numbers across platforms
- Better capture error scenarios in TCP managed SNI.
- Fix logging bug in SqlClientEventSource.
- Change nativeError from uint to int

* - Removed duplicate SniErrorDetails object and aligned field names.

* - Added localized string for the new connection timed out exception.

* - Fixed tests sensitive to OS newlines.

---------


* Align SqlException Numbers across platforms

- Manually fixing some error reporting that wasn't cherr-picked because it was part of larger changes.

* Updated TdsParser.Windows.cs to use the SniErrorDetails struct from TdsParserStateObject.

---------

+++ src/Microsoft.Data.SqlClient/netfx/src/Microsoft/Data/SqlClient/TdsParser.cs
@@ -1586,7 +1586,7 @@ namespace Microsoft.Data.SqlClient
             string providerRid = string.Format("SNI_PN{0}", (int)details.provider);
             string providerName = StringsHelper.GetString(providerRid);
             Debug.Assert(!string.IsNullOrEmpty(providerName), $"invalid providerResourceId '{providerRid}'");
-            uint win32ErrorCode = details.nativeError;
+            int win32ErrorCode = details.nativeError;
 
             if (details.sniError == 0)
             {
@@ -1624,14 +1624,14 @@ namespace Microsoft.Data.SqlClient
                 // If its a LocalDB error, then nativeError actually contains a LocalDB-specific error code, not a win32 error code
                 if (details.sniError == SniErrors.LocalDBErrorCode)
                 {
-                    errorMessage += LocalDbApi.GetLocalDbMessage((int)details.nativeError);
+                    errorMessage += LocalDbApi.GetLocalDbMessage(details.nativeError);
                     win32ErrorCode = 0;
                 }
             }
             errorMessage = string.Format("{0} (provider: {1}, error: {2} - {3})",
                 sqlContextInfo, providerName, (int)details.sniError, errorMessage);
 
-            return new SqlError((int)details.nativeError, 0x00, TdsEnums.FATAL_ERROR_CLASS,
+            return new SqlError(details.nativeError, 0x00, TdsEnums.FATAL_ERROR_CLASS,
                                 _server, errorMessage, details.function, (int)details.lineNumber, win32ErrorCode);
         }
